LIBNFC Compile Good and BLE seems work.

+++ CC2640_NEW/src/examples/simple_peripheral/cc26xx/app/nfcFiles/ble_uart.h
@@ -8,8 +8,8 @@ void    uart_flush_input(bool wait);
 //void    uart_set_speed(serial_port sp, const uint32_t uiPortSpeed);
 //uint32_t uart_get_speed(const serial_port sp);
 
-int     uart_receive(uint8_t *pbtRx, const size_t szRx, void *abort_p, int timeout);
-int     uart_send(const uint8_t *pbtTx, const size_t szTx, int timeout);
+int     uart_receive(uint8_t *pbtRx, const uint16_t szRx, void *abort_p, int timeout);
+int     uart_send(const uint8_t *pbtTx, const uint16_t szTx, int timeout);
 
 //char  **uart_list_ports(void);
